table mismatch fixes
